requirements.txt

transformers
torch
pyqt5
pyttsx3
mss
opencv-python
pyautogui
pynput
requests
aiohttp
openai
speechrecognition
whisper
Pillow
duckduckgo-search
faiss-cpu
edge-tts
numpy
scipy